-fix subnet Update -add data in database -start validation

+++ Core/Services/Subnet/SubnetService.cs
@@ -2,7 +2,9 @@
 using Core.Repositories.User;
 using Core.ViewModels.Subnet;
 using System;
+using System.Linq;
 using System.Net;
+using System.Text;
 
 namespace Core.Services.Subnet
 {
@@ -21,11 +23,21 @@ namespace Core.Services.Subnet
         {
             var subnet = _subnetRepository.Get(userId);
 
+            if (subnet == null)
+            {
+                throw new NullReferenceException();
+            }
+
             return ConvertSubnetToSubnetViewModel(subnet);
         }
 
         public void Create(SubnetForm model)
         {
+            if (!SubnetFormIsValid(model))
+            {
+                throw new ArgumentException();
+            }
+
             var subnet = ConvertSubnetFormToSubnet(model);
 
             _subnetRepository.Add(subnet);
@@ -33,6 +45,11 @@ namespace Core.Services.Subnet
 
         public void Update(SubnetForm model)
         {
+            if (!SubnetFormIsValid(model))
+            {
+                throw new ArgumentException();
+            }
+
             var subnet = ConvertSubnetFormToSubnet(model);
 
             _subnetRepository.Update(subnet);
@@ -40,6 +57,11 @@ namespace Core.Services.Subnet
 
         public void Delete(int userId)
         {
+            if (userId < 1)
+            {
+                throw new ArgumentException();
+            }
+
             _subnetRepository.Delete(userId);
         }
 
@@ -51,13 +73,49 @@ namespace Core.Services.Subnet
             {
                 Id = model.Id,
                 IP = IPAddress.Parse(model.IP),
-                Mask = Convert.ToString(Convert.ToInt32(IPWithMask[1]), 2).PadLeft(8, '0'),
+                Mask = GetMask(IPWithMask[1]),
                 StartOfService = model.StartOfService,
                 EndOfService = model.EndOfService,
                 UserId = model.UserId
             };
 
             return subnetForm;
+        }
+
+        private string GetMask(string mask)
+        {
+            var binaryMask = new StringBuilder();
+            int decimalMask = Convert.ToInt32(mask);
+            for (int i = 1; i <= 32; i++)
+            {
+                if (decimalMask >= i)
+                {
+                    binaryMask.Append("1");
+                }
+                else
+                {
+                    binaryMask.Append("0");
+                }
+
+                if (i % 8 == 0 && i != 32)
+                {
+                    binaryMask.Append(".");
+                }
+            }
+
+            return string.Join('.', binaryMask.ToString().Split(".").Select(x => FromBinary(x)).ToList());
+        }
+
+        private long FromBinary(string input)
+        {
+            long big = 0;
+            foreach (var c in input)
+            {
+                big <<= 1;
+                big += c - '0';
+            }
+
+            return big;
         }
 
         public SubnetViewModel ConvertSubnetToSubnetViewModel(Models.Subnet model)
@@ -74,6 +132,18 @@ namespace Core.Services.Subnet
             };
 
             return subnetForm;
+        }
+
+        private bool SubnetFormIsValid(SubnetForm model)
+        {
+            // TODO: IP Validation
+
+            if (model.StartOfService >= model.EndOfService)
+            {
+                return false;
+            }
+
+            return true;
         }
     }
 }
