+++ Core/Services/User/UserService.cs
@@ -1,31 +1,40 @@
-﻿using Core.Repositories.Subnet;
-using Core.Repositories.User;
+﻿using Core.Repositories.User;
+using Core.Services.Subnet;
 using Core.ViewModels.User;
-using System.Collections.Generic;
-using System.Net;
+using System;
 
 namespace Core.Services.User
 {
     public class UserService : IUserService
     {
         private readonly IUserRepository _userRepository;
-        private readonly ISubnetRepository _subnetRepository;
+        private readonly ISubnetService _subnetService;
 
-        public UserService(IUserRepository userRepository, ISubnetRepository subnetRepository)
+        public UserService(IUserRepository userRepository, ISubnetService subnetService)
         {
             _userRepository = userRepository;
-            _subnetRepository = subnetRepository;
+            _subnetService = subnetService;
         }
 
         public UserViewModel Get(int id)
         {
             var user = _userRepository.Get(id);
 
+            if (user == null)
+            {
+                throw new NullReferenceException();
+            }
+
             return ConvertUserToUserViewModel(user);
         }
 
         public void Create(UserForm model)
         {
+            if (!UserFormIsValid(model))
+            {
+                throw new ArgumentException();
+            }
+
             var user = ConvertUserFormToUser(model);
 
             _userRepository.Add(user);
@@ -33,6 +42,11 @@ namespace Core.Services.User
 
         public void Update(UserForm model)
         {
+            if (!UserFormIsValid(model))
+            {
+                throw new ArgumentException();
+            }
+
             var user = ConvertUserFormToUser(model);
 
             _userRepository.Update(user);
@@ -40,6 +54,11 @@ namespace Core.Services.User
 
         public void Delete(int id)
         {
+            if (id < 1)
+            {
+                throw new ArgumentException();
+            }
+
             _userRepository.Delete(id);
         }
 
@@ -67,17 +86,18 @@ namespace Core.Services.User
                 MiddleName = model.MiddleName,
                 Surname = model.Surname,
                 Age = model.Age,
-                Gender = model.Gender
+                Gender = model.Gender,
+                Subnet = _subnetService.Get(model.Id)
             };
 
-            var subnet = _subnetRepository.Get(model.Id);
-
-            if (subnet != null)
-            {
-                user.SubnetIP = IPAddress.Parse(_subnetRepository.Get(model.Id).IP.ToString());
-            }
-
             return user;
+        }
+
+        private bool UserFormIsValid(UserForm model)
+        {
+            // TODO: UserForm validation
+
+            return true;
         }
     }
 }
